+++ src/constants/types.ts
@@ -26,7 +26,7 @@ export interface AppState {
   typeModal: Type;
   types: Type[];
   addCustomBadge: () => void;
-  addCustomStatus: (status: TStatus) => void;
+  addCustomStatus: (status: string) => void;
   addEncounter: (newLocation: string) => void;
   addGame: (newGame: string) => void;
   addRule: (entry: TRuleEntry) => void;
@@ -47,6 +47,7 @@ export interface AppState {
     moveFour: number
   ) => void;
   changeDupe: () => void;
+  changeLevel: (encounterId: number, increase: boolean) => void;
   changePokemon: (encounterId: number, pokemonId: number) => void;
   changeNickname: (encounterId: number, nickname: string) => void;
   changeRuleset: (rulesetId: string) => void;