+++ src/store.ts
@@ -76,9 +76,14 @@ const useStore = create<AppState>(
             state.customBadges[state.selectedGame?.value] = [0];
           }
         }),
-      addCustomStatus: (status: TStatus) =>
+      addCustomStatus: (status: string) =>
         set((state) => {
-          state.customStatuses.push(status);
+          state.customStatuses.push({
+            icon: null,
+            key: status,
+            value: new Date().getTime(),
+            text: status,
+          });
         }),
       addEncounter: (newLocation: string) =>
         set((state) => {
@@ -156,6 +161,29 @@ const useStore = create<AppState>(
         set((state) => {
           state.duplicates = !state.duplicates;
         }),
+      changeLevel: (encounterId: number, increase: boolean) =>
+        set((state) => {
+          const index = state.games[state.selectedGame?.value].encounters.findIndex((enc) => {
+            return enc.id === encounterId;
+          });
+          if (index !== -1) {
+            const hasDetails = state.games[state.selectedGame?.value].encounters[index].details;
+            const level = state.games[state.selectedGame?.value].encounters[index].details?.level;
+            if (!hasDetails) {
+              state.games[state.selectedGame?.value].encounters[index].details = {
+                id: state.games[state.selectedGame?.value].encounters[index].pokemon,
+                level: 1,
+                moves: [],
+              };
+            } else if (!level) {
+              state.games[state.selectedGame?.value].encounters[index].details.level = 1;
+            } else if (increase) {
+              state.games[state.selectedGame?.value].encounters[index].details.level = level + 1;
+            } else {
+              state.games[state.selectedGame?.value].encounters[index].details.level = level - 1;
+            }
+          }
+        }),
       changeNickname: (encounterId: number, nickname: string) =>
         set((state) => {
           const index = state.games[state.selectedGame?.value].encounters.findIndex((enc) => {
@@ -224,7 +252,14 @@ const useStore = create<AppState>(
         }),
       deleteCustomStatus: (index: number) =>
         set((state) => {
-          // TODO => SHOULD DELETE CUSTOM STATUS FOR ALL ASSOCIATED ENCOUNTERS
+          const customStatusValue = state.customStatuses[index];
+          Object.keys(state.games).forEach((key) => {
+            state.games[key].encounters.forEach((enc) => {
+              if (customStatusValue.value === enc.status?.value) {
+                enc.status = null;
+              }
+            });
+          });
           state.customStatuses.splice(index, 1);
         }),
       deleteEncounter: (encounterId: number) =>
@@ -300,9 +335,17 @@ const useStore = create<AppState>(
           state.games = newAppState.games;
           state.selectedGame = newAppState.selectedGame;
           state.gamesList = newAppState.gamesList;
+          state.darkMode = !!newAppState.darkMode;
+          state.duplicates = !!newAppState.duplicates;
+          state.missing = !!newAppState.missing;
+          state.nicknames = !!newAppState.nicknames;
+          state.showAll = !!newAppState.showAll;
+          state.suggestions = !!newAppState.suggestions;
           if (newAppState.badges) state.badges = newAppState.badges;
           if (newAppState.rules) state.rules = newAppState.rules;
           if (newAppState.team) state.team = newAppState.team;
+          if (newAppState.customBadges) state.customBadges = newAppState.customBadges;
+          if (newAppState.customStatuses) state.customStatuses = newAppState.customStatuses;
         }),
       massImport: (newEncounters: TEncounter[]) =>
         set((state) => {